Refactor timeline components and improve autosave functionality
- Updated CanvasSave component to utilize useCheckModifiedTimeline for unsaved changes detection.
- Simplified state management in CanvasSave by removing unnecessary useEffect hooks.
- Enhanced saveTimeline function to handle validation and return structured results.
- Refactored SidebarContent to dynamically display node data based on selected parameters.
- Improved SidebarFormEdit to include validation messages and handle unsaved changes more effectively.
- Introduced new convertDate utility functions for better date handling.
- Updated types and interfaces for improved type safety and clarity.
- Added a new DashboardLayout component for consistent layout structure.
- Removed deprecated convertDateReadable module and integrated its functionality into convertDate.

+++ src/services/syncTimeline.service.ts
@@ -6,16 +6,16 @@ import { tempErrorHandle } from "@/lib/tempErrorHandle";
 import { returnDataType, timelineNodeType } from "@/types/types";
 
 
-export const syncAndDeleteDBWithLocal = async <T extends { id: string }>(tbName: string, items: T[]) => {
+export const syncAndDeleteDBWithLocal = async <T extends { id: string }>(tbName: string, items: T[], projectUid: string) => {
     
     let nodeIDs = items.map((node: T) => node.id);
 
     if(nodeIDs.length <= 0 && items.length <= 0){
-        nodeIDs = ["__EMPTY_UIDS_"]
+        nodeIDs = []
     }
     
     try {
-        const result = await syncAndDelete(tbName, nodeIDs)
+        const result = await syncAndDelete(tbName, nodeIDs, projectUid)
 
         if(result.status == 200){
             return {
@@ -32,14 +32,25 @@ export const syncAndDeleteDBWithLocal = async <T extends { id: string }>(tbName:
 
 }
 
-export async function syncAndDelete(dbName: string, mappedID: string[]): Promise<returnDataType<{
+interface DeletedRow {
     id: number
-}[]>> {
+}
+
+export async function syncAndDelete(dbName: string, mappedID: string[], projectUid: string): Promise<returnDataType<DeletedRow[]>> {
   
    try {
-      const result = await prisma.$queryRaw`
-                DELETE FROM "${Prisma.raw(dbName)}" WHERE uid NOT IN (${Prisma.join(mappedID)})
+        let result: DeletedRow[] = [];
+    
+        if(mappedID.length > 0){
+          result = await prisma.$queryRaw<DeletedRow[]>`
+                DELETE FROM "${Prisma.raw(dbName)}" WHERE uid NOT IN (${Prisma.join(mappedID)}) AND project_id = (SELECT id FROM "Projects" WHERE uid = ${projectUid})
                 RETURNING id`;
+        }
+
+        if(mappedID.length <= 0){
+          result = await prisma.$queryRaw<DeletedRow[]>`
+                DELETE FROM "${Prisma.raw(dbName)}" WHERE project_id = (SELECT id FROM "Projects" WHERE uid = ${projectUid}) RETURNING id`;
+        }
       
       if(result && Array.isArray(result) && result.length > 0){
           return {
